Remove plenty of class casts in SC and POG
By changing the prototype of IStateRepository.getState() we can get
rid of plenty of class casts at every call of this method.  The change
is safe (although the compiler cannot see it), because each object
is entered in the table with its own state type which matches by
construction its actual type.

Moreover, in case of error in user contributions, a ClassCastException
will be thrown as soon as the result value is used.