Test properly for variant use
The finalization of the Variant module was recomputing whether a
variant was needed by some convergent event.  This is code duplication
and quite fragile, as the information has already been computed in the
event convergence module.

It is much better to store this information in a state info and
retrieve it properly from there.

No functional change.