Do not ask information about a missing variant
Strengthen the IMachineVariantInfo API by preventing callers from
asking for information about a missing variant.  We now throw an
IndexOutOfBounds exception (which does not make much sense currently,
but will when we have lexicographic variants).

This is a slight regression in the published API, but I think we
can leave with it.  The original choice was really too poor (it could
hide unwanted logical errors in callers).