Fix backward compatibility with Rodin 3.4
We need to check that the variant label is the default one in the
special code for backward compatibility for Rodin 3.4.  If the user
has specified a label, then it should be part of the PO name.

This is especially important in the case where a variant would be
removed because of say a typo so that user proofs still match POs.

Add corresponding regression test.