Fix dialog cancellation that were ignored
Sometimes, the cancellation of a dialog was detected because the
associated value was null.  But this only happens when the user
clicks the Cancel button.  If, instead, he presses the Escape key,
then the value is not set to null.

To fix this, we now systematically check the resultCode of the dialog.