Fix #698 Using 'Open With' context menu on PO does not load the PS
Extract the selected PS status from the selection when the editor
gets created so that we can select it as soon as we get the focus.

This allows to work around the bug when the editor gets created.  But
if the editor is already open, then we cannot do anything (except
reimplement the OpenWithMenu class which only considers files).