Fix compatibility issues with div and mod
The AST parser did not completely implement the requirements that all
multiplicative operators shall be compatible with each other. Self
compatibility for div and mod was missing.

Add appropriate tests so that all combinations are now properly tested.

Note: The Rodin 1.0 parser did not fulfill the requirements either.
Things were even worse as one could not mix multiplication, division
and modulo without using parentheses (except for repeated
multiplication). As Rodin 2.0 had started implementing the
specification, it seemed wise to continue in the same vein and add the
missing compatibilities, rather than reverting to the Rodin 1.0
implementation.

The specification is that of the Kernel Mathematical language dated
30 September 2009.