Define a reference target platform
This file defines the reference Target platform for Rodin development.
It contains all installation units that are needed to build Rodin
plug-ins, including test plug-ins.

It is therefore the best way to share this information among Rodin
Core developers.

To use it, open the file with the Target Editor and click `Set as
Active Target Platform` in the top-right corner of the editor.