Fix failing indexer test
On my machine, when running Rodin core tests with Maven, the
testIndexSeveralProjects was failing almost always.

When adding traces the problem seemed to disappear and has never
been reproduced when running the tests directly from Eclipse.

I suspect that this was due to a race condition when deactivating
globally the indexer in the workspace.  Indeed, in that helper code
the indexing job was canceled, but there is no guarantee that it has
actually stopped.  I have added a call to join() to make sure it is
actually terminated, before running the JUnit test.