Skip flaky sequent prover test
The XProverReasonerTests.interrupt fails sometimes because the reasoner
does not fail as expected.

In this test, we run a test reasoner in one thread (through a
FutureTask) which interrupts the calling thread.  It seems that there
is a race condition where the interruption is not always delivered
before the end of the reasoner run (i.e., completion of the FutureTask).

I have tried several approaches to work around that race:
- increase the reasoner timeout,
- make sure that the interruption has been received before returning
  from the reasoner.

However, none of these approaches has given satisfactory results.

I have therefore disabled this test.