Composite Core update site
Add a script for maintaining a Core update site which is composite.
This allows to have one site that have several children sites.

The previous contents of the Core update site has been moved to the
"old" directory in SourceForge FRS.

The compSite.xml script will need to be updated each time a new release
of the Rodin platform is published.

Note: One can add the update site of the underlying Eclipse in order
to make the basic features of Eclipse available for download during
update.