Fix FR #357: Simpler WD for built-in total functions
When we apply a builtin total function, we do not generate a WD
condition anymore for the functional image. This makes WD conditions
much smaller, especially when using a lot of projections.

Note: There are more cases where we could detect that the function
is total, however these cases should be quite rare in practice and non
idiomatic. This is why they are not handled by special additional code.