Tests: improve specialization checker's handling of type expressions
The specialization checker assumes that it is possible to convert back
and forth between type expressions and types without loss of
information. While this works most of the times, there are some
exceptions. For instance, the type expression S ↔ T becomes ℙ(S × T)
when converted in a type: these expressions are equivalent, but not
equal.

This updates the checker to keep the original type expression rather
than convert it in a type and then back in an expression. This causes a
few tests to need a more complete destination type environment.