Fix bug #787: No property tester isEditable exception for EMF editors
When using EMF editors other than the Event-B one, many exceptions were
repeatedly generated because org.eventb.ui.property.isEditable was not
available.

This property tester can now be applied to any implementation of
IEditorInput. It returns false on inputs that are not a FileEditorInput.

This bug and its fix were reported by Colin Snook and Thai Son Hoang.