Fix #786: Invalid thread in Rodin Editor
Method historyNotification is documented as callable from any thread.
The update of the editor actions and handlers may need to be processed
in the UI thread (this is not sure but known to happen sometimes).
To be on the safe side, we always process the notification in the UI
thread.  It does not need to be synchronous, as we do not use the result
of the update.