Fix bug #799: Predicate extension type checking
It was possible to create an extended predicate which had its attribute
typeChecked to true even though calling typeCheck() on it would fail.
It is fixed by calling typeCheck() when creating the extended predicate.

This also fixes a test case that constructed ill-typed extended
predicates (by providing a parameter of type ℤ rather than ℙ(ℤ)): this
error was not detected before this fix.

Later, it would be better to extend the interface of predicate
extensions with a verifyType() method, like expression extensions, but
it will require creating a new interface and deprecating the existing
one.