Add check to avoid ClassCastException
While trying to reproduce bug #788, a slightly different bug has been
found: if the .bps file has been opened with a different editor (e.g.,
Eclipse's default text editor), the editor variable is not null, but it
is not an instance of ProverUI either. Instead of a
NullPointerException, a ClassCastException is thrown. This adds a check
before the cast and, if needed, shows an adequate error message to the
user.