Create new interface IPredicateExtension2
This new interface asks for the implementation of a verifyType()
method which accelerates the creation of extended predicates.

Add tests specific to the old interface.

Note: I did not make the old interface deprecated, because it would
spread everywhere because of the ExtendedPredicate.getExtension()
method.

The other tests will be upgraded to the new interface in the next
commit.