Fix bug #801: Hypothesis checkbox not refreshed when clicked
The bug was caused by setting the background color of the checkbox: with
some desktop environments, it causes the checkmark to not be displayed
anymore when the checkbox is checked. Removing the background color on
the checkbox fixes the bug and does not seem to cause particular issues
with any other system. The predicate row that contains the checkbox
still has its background color around the checkbox.

The dropdown button for tactic application does not suffer from the same
bug, so its background color is not removed.