Fix bug #701: Translation occurs unexpectedly
As suggested in the bug report, the translation now only occurs when the
caret is placed outside the characters of the symbol combo (except when
the widget loses focus: then everything is translated).

As a consequence, many tests had to be adapted to account for this new
rule, typically by changing when the whitespaces are inserted, so that
the cursor ends up outside the symbol combos to be translated.