Display progress dialog for all tactics, interruptible or not
Rodin only showed a progress dialog for interruptible tactics. If a non
interruptible tactic took a long time, there was no indication in the UI
that something was happening. This shows a progress dialog in all cases.
Note, however, that tactics that are expected to take some time should
still be interruptible, so that clicking on the "Cancel" button of the
progress dialog cancels them; otherwise it does nothing.