Deprecate testGetPositions method to test reasoners
Returning an array of values to test and the iterating on it makes it
harder to identify which test case fails. testGetPositions() is now
deprecated. Tests are expected to test positions by creating a new test
method and calling testGetPosition() for each test case. A few test
classes have been updated to do this.