Fix bug #814: ArrayIndexOutOfBounds when typing extended predicates
The fix for bug #799 added a new interface, IPredicateExtension2, and a
workaround to ensure that formulas created with the old extensions are
well typed. This workaround involved creating a list of the bound
identifiers used by the extended predicate before type-checking it.
However, there may be "holes" in the used bound identifiers (e.g., the
children use the first and third bound identifiers, but not the second),
which caused a crash.

It appears that this workaround was needlessly complex. At that point,
we know that children are typed, so we can simply type-check the current
predicate, which does not require to build a list of bound identifiers.