Rework induction to work on a quantified predicate
The induction can now only be applied to the first quantified variable
of a forall at the root of the goal. It is much more restrictive than
before, but it should not be possible to prove false with this.

Since the previous implementation was broken, the tactic is now
versioned.