Add more equivalences between integer inequalities and natural sets
This adds equivalences from inequalities with 1 or -1 and set membership
to NAT or NAT1. The reverse was already implemented.

Some checks are moved to slightly optimize the implementation by only
testing if an expression is the integer zero once we have established
that it is an integer.

More tests are added for these cases and a few older cases that had been
missed.