Instantiate forall with induction value instead of using equality
The previous induction implementation was generating a new hypothesis in
each case (x = C with x the identifier on which induction was done and C
the constructor for the current case). Since we only do induction on a
quantified predicate, we can now easily instantiate the predicate in
each case with the corresponding constructor without generating an
intermediate equality hypothesis.