Fix bug #816: ConcurrentModificationException in editor
Change handling of proofObligations attribute in explorer's models:
either we ensure that we do not modify a map that could be in use, or we
use a ConcurrentHashMap.

There are still other potential race conditions in these models, but the
most common case, as reported in bug #816, should be fixed.