Fix bug #812: dropdown arrow invisible in Event-B editor on macOS
It appears that repeatedly changing the image of the arrows with
setHyperlinkImage(), and therefore repeatedly enabling/disabling the
button, causes the image to gradually fade out on macOS, presumably by
blurring it a bit each time. Calling setImage() directly fixes the bug.

It appears that setHyperlinkImage() was introduced as a workaround for
bug #388 in 2010. It is unclear whether it is still needed nowadays; not
using it for arrows does not seem to trigger the exception reported in
bug #388.