Put back default constructor for backward compatibility
The default constructor of ModelPOContainer was removed when bug #816
was fixed. This should not have been an issue: the class is in an
internal package and all its sub-classes in Rodin Core were updated.
However, it appears that it is used in the Theory plug-in to extend the
explorer. This should be avoided (see FR #381), but meanwhile we put
back a default constructor to avoid breaking the Theory plug-in.