Fix refreshing issue in hypothesis view
An optimization has been added in commit 5fa49a2 to only refresh the
prover's hypotheses views when they are visible. However, it does not
work well and sometimes the views are not refreshed at all. The
setFocus() method which is used to refresh the view is not always called
when the view becomes visible; moreover, even using a focus listener
would not work as the view can be visible without having the focus.
Therefore, it seems better to just remove this optimization and always
refresh the view when the user support manager reports a change.