Allow extension providers to throw core exceptions
In order to avoid changing the API, the interface for extension
providers is not modified, but if they throw a CoreException
encapsulated in the new RuntimeCoreException class, it is caught by the
registry and the CoreException is re-thrown from there.

As a consequence, getting the formula factory of a root can now throw an
exception, which has to be handled. To ease the transition, the
getFormulaFactory() method catches the exception and returns an empty
factory (similar to the previous behaviour), but is now deprecated. A
new "safe" method that throws the exception is added.

The static checker handles the exception by setting a single error on
the root element and stopping there (while, previously, it would try to
run the modules; then, each would generate some errors due to missing
extensions).