Explicitly accept to fall back to default factory in some UI methods
When building content proposal objects, we need a formula factory to
determine if an identifier is valid. If the underlying root object is so
broken that we can't get its formula factory, we consider that it is
acceptable to fall back to a default factory. It may return invalid
results, but at least the UI will keep working and the user can fix them
later, once the root is not broken anymore.