Handle formula factory safely in prover UI
When working on a specific predicate, we use this predicate's factory
rather than the one from the user support (which may have been broken
since).

For the proposal provider, we use the factory from the current PO. It
may be null, but when it happens, the proposal provider returns an empty
list, so it should not need the factory anyway. This requires a small
adaptation to the abstract proposal provider, such that it does not
precompute the prefix (using the factory) before computing the
proposals.