Rework tests about positions
The purpose of this test is to show that positions of possible
applications are correctly computed.  This makes sense only for the
current level, we do not care about past levels.

Moreover, the positions are computed with the same piece of code as the
rewriting, so avoid testing them separately.  This allows removing
plenty of test vectors that we do not need.