Ensure backward compatibility at levels 0 and 1
We restore the previous erroneous behaviour of the reasoner at levels
0 and 1: the auto-rewriter is tried first when rewriting.  Now the
creation of level L2 is justified by removing this erroneous behaviour.

The tests are adapted consequently, by checking that the rules about
comprehension sets are implemented by all levels. Also, the test about
the fact that the auto-rewriter has been removed is adapted to the
different levels.

Finally, method rewriteRootL2 is removed as it is no longer used.