Add notice about the auto-rewriter
In all these classes, the rewriter applies the auto-rewriter first.
This is a mistake. However, it should have no consequence in practice
because the auto-rewriter will not do anything when the reasoner is
applicable. However, during replay, it may happen that the auto-rewriter
does trigger a rewrite if the original predicate has changed.

No functional change.