Fix bug #827: Post-tactic applied after a tactic failure
The API in IUserSupport has been extended to offer a clean way to run
the post tactic in a different thread. Now, we can, in a same call, run
a non interruptible tactic in the UI thread and the post tactic in a
background thread with a progress bar. This is much cleaner and fixes
several issues (the post tactic running after a tactic failure, the
status messages being overwrote, ...).