Improve display of rules
The list of checkboxes is now labeled with the rule names rather than
their IDs and it is sorted alphabetically. The description text box
shows the ID and version (if the reasoner has a version).

While this is clearer than before, it is not perfect. In particular,
reasoners that have several levels all have the same name. So, for
example, there are six checkboxes labeled "Automatic rewrites"; one has
to look at the ID in the description to find out the difference between
them.