Restore proof tree node label filtering
The pattern field was ignored. It is now used to filter out nodes based
on their label, similarly to the filtering based on checked reasoners
checkboxes.

The syntax of the pattern is the same as before (comma separated
patterns that may include * and ?), except that the escape sequence ,,
for , has not been reimplemented: it made everything much more complex
and there does not seem to be much interest in filtering nodes based on
a comma.