Improve display of rules for reasoners with several levels
As noted in commit ee574f6, the list of rules displayed several
reasoners with the same name, one for each level. Reasoners are now
grouped by name. For example, there is now only one checkbox for
"Automatic rewrites" which is used to show or hide all six levels of
this reasoner.