Fix bug #823: Do not allow prime in extension symbol
Actually, primes are not allowed in extension symbols. The bug report
was based on a confusion between identifiers accepted by
isValidIdentifierName (which allows a prime) and checkSymbol (which does
not). Documentation has been updated to clarify the differences between
these methods. Moreover, that the method isValidExtensionSymbol was
added in response to FR #400, checkSymbol does not seem particularly
useful and is therefore deprecated.