Better usage of interrupt attribute
The interrupt attribute should only be true for long-running tactics
that users may want to interrupt and for tactics that need to interact
with the UI. It seems that, at some point, new tactics have started
being all defined with interrupt="true", which is a mistake. All tactics
that have been added in the past few years have been changed to
interrupt="false". For older tactics, it would be too long to sort out
which ones actually need interrupt="true", so they are left as they were
defined.