Do not apply the post tactic twice when running it explicitly
When a user was clicking on the "broom" button to run the post tactic,
the UI would first run the post tactic as requested by the user, then
the automatic post tactic. Naturally, the second post tactic run does
not make sense in that case and does not do anything useful. It is now
disabled.