Clarify when skipPostTactic attribute is used
At first sight, it seems surprising that tactics like prune, show cache,
info, ..., do not define skipPostTactic="true" as it makes no sense to
run the post tactic in such cases. Actually, the post tactic never runs
on tactics that define a proofCommand. It is only run (or not, depending
on skipPostTactic) on tactics that defined a tacticProvider.