Better parsing of quantified expressions
We now use a heuristic to decide between explicit and implicit
quantified expression parsing. This allows the user to see a clearer
error message when the input is wrong.

The heuristic is the same as in Rodin 1.x (with a bug fix: we go to the
next token before checking for a close parenthesis).

Add unit test showing that we now have a readable error message (at
least not worse than in the absence of quantification).