Remove old, unused code in keyboard UI that was noted to be broken
The implementation of translate() in the abstract class was overridden
in LegacyRodinKeyboardTranslator, which is used in Rodin. The other
sub-classes were not used anymore. Therefore, these unused sub-classes
and the broken implementation in the abstract class can be removed, as
well as a dependency that is not used anywhere else.