Fix bug #829: Autorewriter generates invalid proof rule
The set used in rule SIMP_COMPSET_SUBSETEQ now has its bound identifiers
shifted correctly. Since this bug could have generated working but
completely false rules, we have to change the version of all levels
affected, that is L2, L3 and L4 (the rule was not implemented in L0 and
L1, while L5 was added recently and has not yet been released).