Fix another bug related to bound identifiers
While fixing bug #829, it appeared that there was another bug in the
same rule. The rule is supposed to match something like {x · x⊆S ∣ x},
but there was nothing to check that the x was locally bound in that
comprehension set. Therefore, the rule could match and rewrite sets in a
completely wrong way.