Remove TODO that should not be implemented
While experimenting with implementations of the new input class for ae,
we tried to implement this TODO. It appears that several reasoners now
rely on the fact that the input reader throws an exception when a key is
not found. In particular, the exception causes a call to the repair()
method of repairable input reasoners, which then handle the missing key.
Switching now to a null return instead of an exception would probably
cause lots of issues. It's better to leave it as it is.