Remove old update sites that do not work anymore
AnimB is unmaintained and the update site does not work in https.

The update site for Scala is not available anymore. As Scala is only
used by the Isabelle plug-in which last release was in 2012, it does not
seem very useful to hunt down another update site with the right Scala
version.