Put children WDs first in WD of extensions
When computing the WD of an extended formula, the WDs of its children
are now put before its own WD, not after. This is in line with what is
done for all other types of formulas.

A single conjunction with three elements is used, instead of nesting
several conjunctions which would be flattened later on anyway.