Fix bug #832: Bug in translation of datatypes to set theory
The axioms generated for the set constructor of datatypes are changed.
Instead of being defined as a total surjective relation that defines a
partition w.r.t. the constructors, it is now defined as a fix point.

For example, for the usual List(S) inductive datatype, we now generate
List = (λS· ⊤ ∣ (⋂ List ∣ nil ∈ List ∧ cons[S × List] ⊆ List))
instead of the two following axioms:
- List ∈ Message_Type <<->> List_Type and
- ∀T·partition(List[T], {nil}, cons[T × List[T]])