Fix input of CardComparison
The reasoner was inheriting AbstractManualInference which input is a
predicate-position pair even though there was only one valid input: the
goal at the root position. It now inherits EmptyInputReasoner which is a
better match for this. This change is compatible with existing proofs
since empty inputs do not do anything on deserialization.

Tests have been updated to use more modern test methods.