Synthesize type of non-generic datatype constructors
In the case of a non-generic datatype, it is easy to find the type
of constructors, it is the datatype itself. By adding this piece of
code, we can later avoid printing an oftype annotation for these
constructors.

Note that we must nevertheless check the types of the arguments, if any.

Also note that some tests are currently commented out, waiting for the
possibility to add oftype annotations to more operators.