Publish test extensions
Some of the extensions in the Extensions class were not usable
programmatically in tests. This happened because their instances
were created at the same time as the published factory, but not
accessible in any other way.

Now all these extensions are singletons and can be used freely in tests.

No functional change