Add tests about generic extension operators
There were tests for predefined generic atomic operators, but none
about expression extensions.

For the moment, we have only tests about cases where either a type
annotation is not needed, or is added because the operator is atomic.

We will add additional test cases corresponding to bug #834 where a
type annotation is needed for a non-atomic operator later.

No functional change