Start implementing needsTypeAnnotation
We now compute whether a datatype constructor will need a type
annotation to be parsed back with the same type.

In the tests, I have added some type parameters in order to exercise
needsTypeAnnotation. This necessitated to add some more constructor
arguments to satisfy the datatype restrictions in existing tests.

No functional change