Improve toStringWithTypes
Previously, this method would succeed on predefined atomic operators
without a type, but fail for an untyped atomic expression extension.

Now it succeeds in both cases.

Note that the method has to succeed for predefined atomic operators,
otherwise the version upgrader from the V1 language to V2 would fail in
the Event-B core plugin.

Add tests exercising the behaviour.