All datatype parameters must be declared
Enforce a new restriction on datatypes, asking that any given type
occurring in an argument type of a datatype be either the datatype
itself, or a type parameter.

This introduces a possible backward compatibility issue, but should not
cause any problem in practice because this restriction is already
enforced by the theory plugin.

Add corresponding validation tests.