Fix IdentsChecker
We now know that an extended expression can carry a type that is not
fully synthesized from its children, and thus can contain other free
identifiers that need to be recorded.

In IdentsChecker, atomic nodes where not handled in the same way as
non-atomic ones. In the first case, we were taking the type into
account during checking, but the type was ignored for non-atomic nodes.

This is now fixed by always considering the type of any expression node
when present.

This change does not impact existing tests as this case was never
checked. So I have added a new test for this case.

No functional change.