Extend deserialization of ae's input to allow multiple added identifiers
Instead of assuming that ae added a single identifier and searching a
hypothesis matching ident=..., ae's input now accepts multiple added
identifiers and searches for a hypothesis that is an equality which
left-hand-side contains all the added identifiers (and only them).

For the current implementation of ae, it doesn't change anything, but it
will be useful to extend it to accept more complex patterns.