Fix bug #838: Overflow of description in tactic profile editor
First, the description widget must fill its parent in order to adapt to
window resizing.

Second, the 200x200 size should be provided as a hint instead of a hard
minimum requirement, which was causing an overflow in small windows.

This also happens to fix overflows in the list viewers of the tactic
editor: when the window is very narrow, these viewers now get a
horizontal scroll bar.